Updated openstack/openstack Project: openstack/nova  0ecb10c6502cff8def9160f24d31e2af9baaf573
Remove 'scheduled_at' - DB cleanup

The 'scheduled_at' field is a part of the Nova instances.
Currently, the 'scheduled_at' field is no longer updated
by the Nova scheduler, instead the 'launched_at' field is
updated for the same purpose. The 'scheduled_at' column is
now redundant and should be removed as a part of DB clean up.

Implements: blueprint cleanup-scheduled-at
Change-Id: I66b55bbd4ad7c381e6a7203c22ac7c2d29e222f6