Updated openstack/openstack Project: openstack/tempest  8a5a3e7582040c63681de6164ac33558bc5b2499
Include sample configuration in dev docs

Since including the sample configuration in tree is
becoming difficult to maintain, lets try moving it to
the developer docs. We still have the ability to view it
easily for users, and it can be auto-generated on doc build,
so it should also always be up to date.

Change-Id: I7ec2d76448feef5ecb9af7c238bbd12619fa31d8