Updated openstack/openstack Project: openstack/python-keystoneclient  26534dadb1d0be00b87b632a038839ab1c18cfe4
oslo-incubator apiclient.exceptions to keystoneclient.exceptions

Applications are using exceptions out of
keystoneclient.openstack.common.apiclient.exceptions so it's part
of the public interface. But since it's from oslo-incubator it's
not considered stable. Since keystoneclient is all stable this
creates bad situation.

With this change, all the symbols out of apiclient.exceptions are
moved into keystoneclient.exceptions rather than the other way
around (keystoneclient.exceptions used to re-export all of
apiclient.exceptions). Now we're in control of the
apiclient.exceptions and keystoneclient.exceptions that
applications are using.

Closes-Bug: 1481806
Change-Id: Ib3afa86b9d276f6a45d1ecd6f9157ee02ec8570d