Updated openstack/openstack Project: openstack/cliff  0a8daff7bdad7c56bc7b9dc2d6dea6b7259d2827
Implement a yaml formatter in cliff

This change replaces the cliff-tablib yaml formatter with an internal
replacement. It differs from the tablib formatter in the following ways:
- always outputs with block formatting rather than the PyYAML default
  of deciding based on value content
- emit_one serialises a simple dict where the column name is the key
  and the data item is the value (rather than a list of dicts with
  'Field' and 'Value' keys)
- emit_one preserves column order by printing a single-item dict for
  each column_name

The cliff release which contains this change will need a corresponding
cliff-tablib release which removes the yaml formatter from its setup.py
entry_points.

Change-Id: I691dbab3dee7c5ec28b1083f87ab1f5c051d582b
Related-Bug: #1308744