Updated openstack/openstack Project: openstack-infra/project-config  0c9076ced867dc6b92182bb803e0963c531a2976
Add separate requirements validation job

We used to validate the requirements checks through the py27 unittest
job, but this resulted in output that was hard for contributors to
understand. Those tests have been moved out to a separate command, with
its own tox environment, so set up a job to run the new tool.

Depends-On: Icb717b0f36afb6ea29f50bc6935917dddf47fd4c

Change-Id: Ib1bd44cf710b89101f28b0a9db5ede473e5e8c25