Updated openstack/openstack Project: openstack/horizon  ca4a38c345ef8d7c6f23d21e7fdb7706ff7381ed
Use setUp() method for neutron rest test

There is no reason to override the setUpClass() in the test
case. For data setup, the setUp() should be used.

There is also an issue where the code is not calling super()
class, this issue is exposed when running the test in django18.

Closes-Bug: #1481143
Partially Implements: blueprint django18

Change-Id: Ic64cc0a5f0c7d0476bfac63b6e23fe8a1461d6ef