Updated openstack/openstack Project: openstack/ceilometer  6ddee05da8923bad3182110d8d80e2b10f772540
gnocchi: retry with a new token on 401

The gnocchi dispatcher just bail out on any unexpected Gnocchi HTTP code.

This change moves all Gnocchi API and HTTP stuffs in the gnocchi_client
module.
And improves the handling of 401 error code to retry requests when the
authentication fail, usualy because the token have expired.

Change-Id: I9d7e29337de2cb480c64af03032dcf9fa2a3829c