Updated openstack/openstack Project: openstack/nova  91f8cc9c153b61a5aed081c2d1b44b21f35d3311
cleanup NovaObjectDictCompat from virtual_interface

cleanup subclassing on NovaObjectDictCompat and fix subsequent
tests and code associated with nova/objects/virtual_interface.py

Due to the exhaustive nature of changes, the cleanup is done
one object at a time.

This is the second patch of the series.
There shall be more patches to follow for other objects.

Related to blueprint liberty-objects

Change-Id: Ia66bea59828beaeb8d79348331619d0b848c1a3f