Updated openstack/openstack Project: openstack-infra/shade  b55f88bddc211b46fe6513ae11c2b743469a1dd6
Be consistent with accessing server dict

We had several different places where we were treating the server
dict as an object. Although it can currently be both because we use
Bunch objects for backward compatibility, we should be consistent
with treating it as a dict so that we can, hopefully, at some point
stop using Bunch objects. And also we don't want to continue to
provide bad examples.

Change-Id: I27eb467ac2e86454a86566336a55e892e7025fda