Updated openstack/openstack Project: openstack/oslo.db  35a93a235d25b9daa3470db5c825893e8f73c1a7
Upgrade and downgrade based on revision existence

Before there was no way to upgrade to a specific revision, because the
revision passed to the manager was passed to all the plugins, some of
which failed to process it.

Add a new method to every plugin `has_revision`, which returns
whether the plugin has the revision. Use it in upgrade and
downgrade.

Change-Id: I89b02d7ad479da6bff3c492c88edfee9c19abc22
Closes-Bug: 1474067