Updated openstack/openstack Project: openstack-infra/project-config  ce443ff86bfcd1917ea4ff75ca9d27449f669da3
Fix search for docs tox job in run-docs.sh

The run-docs.sh script was checking for a tox job using the tox
-l/--listenvs flag. However this only shows jobs that are listed in
the envlist variable in the tox.ini. The docs job is not normally put
in this variable because it's what is used to define the default set
of jobs run when you just run 'tox' without any args. This commit
changes the if check to search the entire tox config for any jobs
listed in the configuration labeled docs. This should pick up docs
tox jobs more reliably.

Change-Id: Ieeccba08fe7960206864723e6f13a10309cfacf3