Updated openstack/openstack Project: openstack-infra/infra-manual  21f4c3a3a3da3e97db8ac69582bb191d8af65d87
Expand on directions for creating drivers team

We have quite a few drivers teams where the openstack-release team is
not a subteam, so the release team has no permissions to create or
update milestones in launchpad. Clarify the instructions here so we can
refer to them when telling projects how to fix their configurations.

Change-Id: Icf5ef4b71c05450be2e5c67a93e217db7afa759a