Updated openstack/openstack Project: openstack-infra/system-config  065fde589b4b3db89133cb12dccffb7fcc59d77a
Don't clone project-config for puppet-apply jobs

It does not appear these jobs actually require the project-config repo
for anything.  It was added in
I3a772e84a1bc6028caeabcc8e09a0942f44b8612 without a lot of
explaination of why, but as it stands I don't see the project-config
affects the outcome of these jobs?

Change-Id: I85111afbbce497c86f4f6d8dfe74944c32a77b80