Updated openstack/openstack Project: openstack-infra/project-config  40bb9094a1e447ae9e4f9f3188d0e5d0d69c712e
Don't run puppet-apply jobs for project-config

My first thought was that we don't need to be running the puppet-apply
tests for the majority of changes that come into project-config that
just add or fiddle job configs.

But then I thought that if the puppet-apply jobs aren't relying on
project-config (see dependent change) then I don't think running them
here helps.

Change-Id: I2e24b300fe31e6eec5acaca48fba94cd5633f38f
Depends-On: I85111afbbce497c86f4f6d8dfe74944c32a77b80