Updated openstack/openstack Project: openstack/oslo.versionedobjects  5be082c3fec6843dc44cd1d10a1548b5746f2009
Add obj_tree_get_versions() utility method

This adds a utility method that takes a top-level object and determines
all of the dependent object names and versions that could be nested
within. This can be later used to ask the indirection_api to backport
an object with this information instead of depending on its own mapping,
which may differ.

Change-Id: I04574ccd01e582b3eef8f98cac41f429c70debd8