Updated openstack/openstack Project: openstack-infra/zuul  fe707d1599727f1dacfcc7801d088915609d1303
Fix bug with shared jobs in reconfiguration

This is a variant on the bug fixed in
I2d5cc8f4c5b73d6a2b184cd21016943b8c08949d.
This corrects the case where a job is removed from a project and
there is a current failed build of that job for that project in the
queue, yet, unlike the case in the other bugfix, this job still
exists elsewhere in the system.

Now we verify that a job exists at all, and also that it is still
supposed to be run for the change in question when deciding whether
to keep it in the build-set for the change.

Change-Id: I0829d42612e7e74d789c60d0570e5bdf9ec47937