Updated openstack/openstack Project: openstack/oslo.versionedobjects  a05639ba2a615b7b76e94505c6246797720776d9
Implement multi-version object backport method

This adds a new backport method to the VersionedObjectSerializer
interface. The new method calculates a manifest of all the local-side
object versions expected to be needed during the backport request and
sends that to the server. Ideally this will replace the current
backport method that depends on the version relationship maps that
we maintain, and will limit the complexity involved with relationship
maps differing on older code.

Change-Id: I071c20b65d12df893c367d417ab6be09626a9955