Updated openstack/openstack Project: openstack/oslo.middleware  3fbb59574ca3bceb42baa0847d146f471145ab31
Added latent properties to CORS middleware.

Latent properties allow a consumer of this middleware to declare
system-required headers and methods options. For instance, if an
API exposes version-negotiation headers, these may be hard coded
when the middleware is attached.

This only works when the middleware is explicitly used. It does
not work in paste configuration.

Change-Id: Ic55b1af23603a0d83a32d20054c18e50367be8fb