Updated openstack/openstack Project: openstack/nova  7192432acc845ddff1c36a3ec91d7eaf4aa31731
libvirt: move LibvirtNETVolumeDriver into it's own module

Introduces the LibvirtISCSIVolumeBaseTestCase class for the iSCSI-based
volume driver tests like iSCSI and the Network volume driver (iSCSI volume
driver changes yet to come).

There is some additional refactor cleanup that can be done here since
this code is a bit hacky with the protocol checks and secret CRUD ops in
the libvirt host module, but that's reserved for later changes.

Part of blueprint consolidate-libvirt-fs-volume-drivers

Change-Id: I31404d057d596473a9c0960e2696affb1e35e132