Updated openstack/openstack Project: openstack-infra/project-config  42fc124e4e6e7edd6e0adcad1907561837a9a682
Add better log + console extraction to anvil-jobs

To be able to extract the rpmbuilding logs that typically
are created add a tweaked job that should extract these and place
them in a location for external viewing (helping in any debug
that is needed when it breaks).

This also ensures that the openstack directory where everything
is built and checked-out is in the local directory by explicitly
using the '-d' option to ensure the jenkins home directory isn't
the target instead (which will be the default if not specified).

Change-Id: Ib0bbd72248e73baaf66c8acf9984b69d119c13df