Updated openstack/openstack Project: openstack-infra/project-config  e432c1c48c9ca65f0de976fdfb7ba2693d4f8b60
Adding swift3 to list of projects for particular ec2 job

Additional project should be added to PROJECTS as I understood.
It needs for clone the project to /opt/stack if this project enabled
for devstack.

Also, as Ian Wienand asked about regrouping definitions in patchset #1.
So I regrouped definitions in another job too to be similar.

Change-Id: If7e891cdedb3342a326253bca7391a0495959f69