Updated openstack/openstack Project: openstack/neutron  3d5eab8fa796d4ffc9b769fab4c1736eea4d3647
Move away nested transaction from _ensure_default_security_group

This patch remove the nested transaction started in
_ensure_default_security_group, before calling create_security_group.
Instead, a nested transaction will be started inside
create_security_group.

The purpose of this change is to reduce the time period in which the db
transaction is open (to avoid races, db lock timeouts, etc), when creating
the default security-group.

Closes-Bug: #1479558
Change-Id: Ia0efa72c70c2f405c851370b0b26770008e5ff95