Updated openstack/openstack Project: openstack/ceilometer  392fc100f445abf788154db19425e2b78b872780
Sync devstack plugin with devstack:lib/ceilometer

While the removal of ceilometer from core devstack is still pending
we need to make sure that change in lib/ceilometer are reflected in
the plugin. This synchronizes the small number of recent changes.
We're nearly there on the extractions so there shouldn't be too much
more of this.

Change-Id: I8c99d57ba5183d3c63b4814ca0664d5bf3420832