Updated openstack/openstack Project: openstack/tempest  26c0fabfb1169c6d494bd9a267dc6a8b4c3cd491
increase failure information for ip_regex test

It's basically completely impossible to understand why this test would
fail because there is not enough post mortem information. Probably an
address didn't allocate, but this is so stripped down you'd never know
that.

Add further dump on failure to make it possible to see what's happening.

Change-Id: I536f78c77df992dde5efd22b1d763715e33f0769
Related-Bug: #1481715