Updated openstack/openstack Project: openstack-infra/elastic-recheck  95f5d361e0fa57682fe1e0f5eaed73e8891e2ada
Track nova net dhcp failures

There are cases where nova boots nodes and they fail to get an address
because dnsmasq claims there is no address available. This appears to be
a race between the node starting and doing a DHCPDISCOVER and nova
network configuring dnsmasq.

We see this on both successful and failed jobs because tempest does not
always attempt to ssh into the nodes that it boots. In these cases
nothing checks that dhcp worked allowing for jobs to pass.

Change-Id: I85cea28f905ee8fbe769b53026595c83e8cbfdde
related-bug: 1445569