Updated openstack/openstack Project: openstack/api-site  bc440c0f4d72218ef43d1c1d19d48022d9d54011
Add "network" and "networks" objects to Network v2.0 extensions

"Request codes" and "response codes" tables of these
sections are missing "network(s)" (xsd:dict).

  - Networks provider extended attributes
  - Networks multiple provider extension
  - VLAN transparency extension

This patch adds it to all the tables. These sections
refer to the same ENTITYs, so fixing the ENTITYs
fixes all the sections. Having said that, I don't think
it's a good idea that all list, create, show, and update
operations use the same ENTITY, which is "networkListParameters"
since that makes it harder to maintain the file. I split
the ENTITY for each method individually, introducing a
common ENTITY that other ENTITYs can refer to.

  - networkCommonResponseParams
  - networkListResponseParams
  - networkCreateReponseParams
  - networkShowResponseParams
  - networkUpdateResponseParams

Change-Id: Ib969102eb509bf1096bb48f865885ed9542b84c8
Closes-Bug: #1480312