Updated openstack/openstack Project: openstack/nova  bfec231f80bec77fed272dc09f2660656f4c2675
Add test to check relations at or below current

In change I8e2a9bd3bfa049bc099a15df3a14ad84e35ac372 there was a list
object that had its relationships updated, but did not have its own
version updated to use the new relationship.

This change adds 2 tests: one to ensure the parent versions in
obj_relationships are not past the current version of the parent, and
one to ensure that the child object versions in obj_relationships of all
objects are not past the current version of the child object.

Also updated a version to pass the added unit tests (and to fix the
object itself).

Change-Id: If9189c57290dcb19a0efd94bcfa02f9c4eef53d1