Updated openstack/openstack Project: openstack-infra/system-config  53676ea705fd49552d2215bf98ee39773e54c3d7
Added NPM Publishing credentials to pypi slave.

This patch adds an .npmrc file for the jenkins user on our pypi slave.
Note that actually publishing to npm will only work on trusty or later.
This patch requires a manual step, where a member of infra-root registers
the openstackci user on npm, and injects the password into hiera.
Note that npm credentials require a user, password, and URL, some of which
are specified in site.pp.

Change-Id: I13fa91500a7b189e5fb6e915968727e8c1f2a1f4