Updated openstack/openstack Project: openstack-infra/nodepool  7a86e46d313dc41d197be707560f5f50dd7e7391
Shorten template hostname

Previously we were setting template hostnames to a very long name that
included the domain making it an fqdn that does not resolve. This causes
problems when that hostname is truncated by the system but other places
try to use the full long name.

Work around this by dropping the template.openstack.org default domain
attached to the hostname and change the default template hostname to
'template-{image.name}-{timestamp}'. This should keep names to a
reasonable length allowing hostname related operations to function
properly.

Change-Id: Ic55f018d31646975af88d38d2df3eecd72b13f8b