Updated openstack/openstack Project: openstack/neutron  453d20996b8babec32d041f0698a3a08f4386014
NSX: Rename default_interface_name option

Since this option is configured in the DEFAULT section of nsx.ini,
the name of the option clashes with that in the networking_l2gw repo.
Proposal is to prefix our option with "nsx_" to avoid such interference.
Unfortunately this module was not moved to openstack/vmware-nsx
during decomposition and therefore this patch is being proposed on
the neutron repository.

Change-Id: Ib568333a5ab8ad76600beb4981d7c6047c0603a0
Partial-bug: #1480485