Updated openstack/openstack Project: openstack-infra/nodepool  fb0679ca3f4d26c8d0329dd9a70d5ef5013860d2
Fix erroneously deleting non-leaked instances

The leaked-instance detection runs in a single database transaction
for its entire run, which can be quite long as it involves listing
all servers in all providers.  This means that a database snapshot
is created at the first read, and used for the rest of the process.
Meanwhile, the instance listing portion of the routine operates as
a series of snapshots, in that each provider is listed in turn.
The database and instance list will be approximately in sync for the
first provider, but as we move through the list of providers, by
the end we will be operating on a much more current list of instances
than our list of nodes from the database.

To correct this, create a new session (and therefore transaction)
after each instance listing.  This will ensure that every nodepool
instance that the provider reports will be in the current database
snapshot because the node record is committed before server creation.
There is a potential exception to this if the node launcher or some
other portion of nodepool has already managed to delete both the
server and database record between the time of the instance listing
and database lookup in the leak check.  However, in that case, we
will simply treat an already deleted instance as a leaked instance
and try to delete it again which is harmless.

Also correct some log lines to be more consistent, and change the
default mysql engine to innodb so that developers running this locally
end up with tables created correctly.  This should also improve
the accuracy of CI tests.

Change-Id: Iadd4c03510332a7bf99a7309a20a292dd5dccf18