Updated openstack/openstack Project: openstack-infra/subunit2sql  4d04042bd64a6c77bf89e9b51ef81fa7d27b3724
Fix issues in migration 1ff737bef438

This commit fixes a couple of issues with the migration 1ff737bef438.
First, on multi-byte configured mysql systems the status column is too
big to index, so this adds a length to the index to allow it to run on
those envs. The second is that since the migration fails to run on
certain envs checks need to be done to account for migrations that are
only partially applied. The min version of SQLAlchemy is also bumped
to version 0.8.2 because that's when using dictionaries on the
mysql_length for composite indexes was added.

Change-Id: Iad056d44251ad2eac8a5cb5628779b27c2126630