Updated openstack/openstack Project: openstack/openstack-manuals  4e177918dae3cabc0b3f68965a3ad52f9af3baa3
Link to Euca2ools Guide goes to Page Not Found

In the Cloud Administrator Guide under the category EC2 API Compatability,
the external reference link to Euca2ools is broken, pointing to HP's site
that returns a Page Not Found error.

This bug fix replaces the broken link with a correct one that points to the
Euca2ools guide.

Change-Id: I5ecfac00c3f0fd4bb59006c37d675fe0bbfa9db2
Closes-Bug: #1480729