Updated openstack/openstack Project: openstack/heat  3f5b8ee9cc84f1f148d17604d0df64784a2ce121
Support boolean inputs for SoftwareConfig's

Adds support for using input/output booleans when creating
SoftwareConfigs. This is useful because some of the software
config hooks that prefer explicit boolean types for input's.
Using Puppet with Hiera to map inputs directly to class parameters
is an example that would benefit from this capability.

Change-Id: Ic80f35fb259e60429d5a65a1b0bc7ff9f2a3ea0a
Closes-bug: #1482223