Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  4e479a8256d1d38c9a534868c4af7b34601a284d
Adds onlyIfSuccessful to ArtifactArchiver publisher

In the Post-build Actions section there is a checkbox for "Archive artifacts
only if build is successful" that does not appear to be affected by the
latest-only yaml key, in Jenkins 1.620, that generates the latestOnly tag.

This patch adds a new yaml key for if-only-success that generates a tag for
onlyIfSuccessful that influences the mentioned checkbox.

Change-Id: I86c50d5c7a9e2234ae9cf4b3b59de7292d1dbaa0