Updated openstack/openstack Project: openstack/nova  bccd17846176646e1503259c4a08e73fa72a4528
libvirt: don't open connection in driver constructor

Calling host.has_min_version triggers opening of a new
libvirt connection. This should not ever be done in the
constructor of LibvirtDriver class, as the 'host_init'
method must be allowed to run first during normal Nova
compute service startup.

This mistake was introduced in

  commit 9a09674220a071e51fdca7911b52c0027c01ff64
  Author: Vladik Romanovsky <vromanso@redhat.com>
  Date:   Wed Oct 15 14:42:37 2014 -0400

    libvirt: enable virtio-net multiqueue

and is thought to also be causing instability in the
unit tests in the gate.

Avoid the problem by passing the Host object instance
into the VIF driver get_config method, instead of
trying to pass a flag into its constructor.

Closes-Bug: #1482230
Change-Id: Ic940d6ffc5b7d664540363b8bcbea23c7bc7d9dd