Updated openstack/openstack Project: openstack/oslo.log  54bc7f2c925d7e66818a98b45622c68d15711fe8
Provide a way to register versionutils options

The versionutils functions will register their options when they're
first called, but this forces applications to call the function in
order to be able to override the option.

A new function is added to register the options independently of
calling the existing functions.

Change-Id: I0f3f4cf0f42fefcb9773a00c1437380b9fd8f21b
Closes-Bug: 1474069