Updated openstack/openstack Project: openstack/nova  6e5cc92c41ba3946577200d1f1cb23214ac464ab
cleanup NovaObjectDictCompat from virt_cpu_topology

cleanup subclassing on NovaObjectDictCompat and fix subsequent
tests and code associated with nova/objects/virt_cpu_topology.py

Due to the exhaustive nature of changes, the cleanup is done
one object at a time.

This is the fourth patch of the series.
There shall be more patches to follow for other objects.

Related to blueprint liberty-objects

Change-Id: I6c4ed28faf0bccf169faec295b410e764f70db5c