Updated openstack/openstack Project: openstack/ceilometer  00aa6cec3ec0c9725649ba3eaed9545a4076a28a
Make ConnectionRetryTest more reliable

The previous test relied on mocks of retrying.time.sleep. This is
the same sleep as time.sleep so there's a chance, if eventlet is in
context, that something else can call that.

So instead the test now checks to see how many times should_reject
is called. This is the same as the max retries that will be
attempted.

Also the retry wait time has been set to something very small so the
test doesn't hang around sleeping uselessly, making test runs
faster.

Change-Id: Ib15d4d6f8dd23ba534d58005da5dc64aad9e0e90
Closes-Bug: #1482722