Updated openstack/openstack Project: openstack/nova  acdb72bfac29607247402e71b031697476c7c137
Implement compare-and-swap for instance update

This patch reworks nova.db.sqlalchemy.api._instance_update to remove a
race condition that exists in a critical section of code that
retrieves the instance record from the database, does some checks of
both the new values and existing values of instance fields, and then
tries to update the database with the new values.

We update the value using update_on_match, which does an optimistic,
atomic, lock-free, compare-and-swap on the object. If the update
fails, we do some additional checks to determine the specific error,
which maintains the previous behaviour.

_instance_update() is simplified by having it return only the updated
instance_ref. instance_update_and_get_original() is updated to fetch
the old value itself before calling _instance_update().

This patch was originally submitted as change
I9cd0f4b620e639b238555983bf6d58deafbaefeb, but that was reverted. The
original version was optimistic about not racing in
instance_update_and_get_original between fetching an instance and
updating it, in which case it raised a different exception to the
previous behaviour. In practice we were hit that race frequently in
the gate, and the unexpectedly different exception was causing gate
failures. This new version adds a retry in this case, and raises the
expected exception. The frequency of hitting that race highlights the
importance of this patch, as it is otherwise unhandled.

This new version also updates the exception hierarchy so that all
exceptions raised by _instance_update inherit from a single exception:
InstanceUpdateConflict. UnexpectedVMStateError is removed, as it had no
users. An InstanceUpdateConflict is now raised instead.

The new code trivially allows the caller to check any property when
updating an instance, not just the two prescribed. This new version
enables this by allowing the 'expected' dict to be passed in, as well
as maintaining the behaviour of pulling expected values from the
update dict. This has numerous potential applications, for example
atomically changing the host of an instance during evacuation.

The new version adds numerous additional tests.

Closes-bug: #1297375
Change-Id: I293da6f320dd8f3474ce2a9c907298e1fb348181